+++ bot/skills/towel_mode.py
@@ -1,160 +1,164 @@
-import datetime
 import logging
+from datetime import datetime, timedelta
+from random import choice
+from typing import Dict
 
-from telegram import Update, User, Bot, InlineKeyboardButton, InlineKeyboardMarkup
-from telegram.ext import run_async, MessageHandler, Filters, Updater, CallbackContext, CallbackQueryHandler
+from pymongo.collection import Collection
+from telegram import Update, User, InlineKeyboardMarkup, InlineKeyboardButton
+from telegram.ext import Updater, MessageHandler, Filters, CallbackContext, run_async, CallbackQueryHandler
 
 from config import get_config
+from db.mongo import get_db
+from mode import Mode
+
+MAGIC_NUMBER = "42"
+QUARANTINE_TIME = 60
+I_AM_BOT = [
+    "I am a bot!",
+    "Я бот!",
+    "私はボットです！",
+    "Ma olen bot!",
+    "मैं एक बॉट हूँ!",
+    "Je suis un bot!",
+    "Unë jam një bot!",
+    "أنا بوت!",
+    "אני בוט!",
+    "Sono un robot!",
+    "我是機器人！"
+]
 
 logger = logging.getLogger(__name__)
 
-QUARANTINE_STORE_KEY = "towel_mode__quarantine"
-QUARANTINE_MIN_STORE_KEY = "towel_mode__quarantine_min"
-MAGIC_NUMBER = "42"
 
-conf = get_config()
+# todo: extract maybe?
+class DB:
+    def __init__(self, db_name: str):
+        self._coll: Collection = get_db(db_name).quarantine
+
+    def add_user(self, user_id: str):
+        return self._coll.insert_one({
+            "_id": user_id,
+            "datetime": datetime.now() + timedelta(minutes=QUARANTINE_TIME)
+        }) if self.find_user(user_id) is None else None
+
+    def find_user(self, user_id: str):
+        return self._coll.find_one({"_id": user_id})
+
+    def find_all_users(self):
+        return self._coll.find({})
+
+    def delete_user(self, user_id: str):
+        return self._coll.delete_one({"_id": user_id})
 
 
-# todo: make it by Mode decorator
+db = DB("towel_mode")
+mode = Mode(mode_name="towel_mode", default=True)
 
+
+@mode.add
 def add_towel_mode(upd: Updater, handlers_group: int):
     logger.info("registering towel-mode handlers")
     dp = upd.dispatcher
 
     # catch all new users and drop the towel
+    dp.add_handler(MessageHandler(Filters.status_update.new_chat_members, catch_new_user),
+                   handlers_group)
+
+    # check for reply or remove messages
     dp.add_handler(MessageHandler(
-        Filters.status_update.new_chat_members, put_new_users_in_quarantine),
-        handlers_group
+        Filters.group & ~Filters.status_update, catch_reply)
     )
 
-    # remove messages from users from quarantine
-    dp.add_handler(MessageHandler(
-        Filters.group & ~Filters.status_update, check_for_reply),
-        handlers_group
-    )
+    # "i am a bot button"
+    dp.add_handler(CallbackQueryHandler(i_am_a_bot_btn))
 
-    # catch "I am not a bot" button press
-    dp.add_handler(CallbackQueryHandler(butt_press))
-
-    # kick all not replied users (check every minutes)
-    upd.job_queue.run_repeating(callback_minute, interval=60, first=60, context={
-        "chat_name": conf["GROUP_CHAT_ID"]
+    # ban quarantine users, if time is gone
+    upd.job_queue.run_repeating(ban_user, interval=60, first=60, context={
+        "chat_id": get_config()["GROUP_CHAT_ID"]
     })
 
 
-def get_quarantine_and_quarantine_min(context: CallbackContext):
-    logger.debug("get quarantine and quarantine_min from context.chat_data")
-    if QUARANTINE_STORE_KEY not in context.chat_data:
-        context.chat_data[QUARANTINE_STORE_KEY] = {}
-
-    if QUARANTINE_MIN_STORE_KEY not in context.chat_data:
-        context.chat_data[QUARANTINE_MIN_STORE_KEY] = [[] for _ in range(60)]
-
-    return context.chat_data[QUARANTINE_STORE_KEY], context.chat_data[QUARANTINE_MIN_STORE_KEY]
-
-
 @run_async
-def put_new_users_in_quarantine(update: Update, context: CallbackContext):
-    for user in update.message.new_chat_members:
-        quarantine_user(user, update.effective_chat.id, context)
-
-
-@run_async
-def quarantine_user(user: User, chat_id: str, context: CallbackContext) -> None:
-    """ Show welcome msg and put user into quarantine """
+def quarantine_user(user: User, chat_id: str, context: CallbackContext):
     logger.info(f"put {user} in quarantine")
-    quarantine, quarantine_min = get_quarantine_and_quarantine_min(context)
-    minute = datetime.datetime.utcnow().minute
-    quarantine[user.id] = minute
-    quarantine_min[minute].append(user.id)
+    db.add_user(user.id)
 
     markup = InlineKeyboardMarkup([
-        [InlineKeyboardButton("Я не бот!", callback_data=MAGIC_NUMBER)]])
+        [InlineKeyboardButton(choice(I_AM_BOT), callback_data=MAGIC_NUMBER)]])
     context.bot.send_message(
         chat_id,
-        f"{user.name} Нажми кнопку ниже, чтобы доказать, что ты не бот.\n"
+        f"{user.name} НЕ нажимай на кнопку ниже, чтобы доказать, что ты не бот.\n"
+        "Просто ответь (replay) на это сообщение что-нибудь."
         "Я буду удалять твои сообщения, пока ты не сделаешь это.\n"
-        "А коли не сделаешь, через час выкину из чата.\n"
+        f"А коли не сделаешь, через {QUARANTINE_TIME} минут выкину из чата.\n"
         "Ничего личного, просто боты одолели.\n",
         reply_markup=markup
     )
 
 
 @run_async
-def check_for_reply(update: Update, context: CallbackContext):
-    quarantine, quarantine_min = get_quarantine_and_quarantine_min(context)
-    minute = quarantine.get(update.effective_user.id)
-    if minute is None:
-        return
-
-    logger.debug(
-        f"delete msg from user: {update.effective_user} [quarantine]")
-    context.bot.delete_message(
-        update.effective_chat.id,
-        update.effective_message.message_id, 10
-    )
+def catch_new_user(update: Update, context: CallbackContext):
+    for user in update.message.new_chat_members:
+        quarantine_user(user, update.effective_chat.id, context)
 
 
 @run_async
-def butt_press(update: Update, context: CallbackContext):
-    quarantine, quarantine_min = get_quarantine_and_quarantine_min(context)
-    minute = quarantine.get(update.effective_user.id)
-    if minute is None:
+def catch_reply(update: Update, context: CallbackContext):
+    # todo: cache it
+    user_id = update.effective_user.id
+    user = db.find_user(user_id)
+
+    if user is None:
         return
+
+    if update.effective_message.reply_to_message is not None and \
+            update.effective_message.reply_to_message.from_user.id == context.bot.get_me().id:
+        db.delete_user(user_id=user['_id'])
+        update.message.reply_text("Welcome to VLDC!")
+    else:
+        context.bot.delete_message(
+            update.effective_chat.id,
+            update.effective_message.message_id, 10
+        )
+
+
+@run_async
+def quarantine_filter(update: Update, context: CallbackContext):
+    user_id = update.effective_user.id
+    # todo: cache it
+    user = db.find_user(user_id)
+    # if user exist -> remove message
+    if user is not None:
+        context.bot.delete_message(
+            update.effective_chat.id,
+            update.effective_message.message_id, 10
+        )
+
+
+@run_async
+def i_am_a_bot_btn(update: Update, context: CallbackContext):
     chat_id = update.effective_chat.id
     user = update.effective_user
     query = update.callback_query
     if query.data == MAGIC_NUMBER:
-        logger.info(
-            f"remove user: {user.name} from quarantine")
-        del (quarantine[user.id])
-        quarantine_min[minute].remove(user.id)
         context.bot.send_message(
-            chat_id, f"{user.name} Добро пожаловать в VLDC! 😼")
+            chat_id, f"{user.name}, попробуй прочитать мое сообщение внимательней :3")
+
+
+def _is_time_gone(user: Dict) -> bool:
+    return user["datetime"] < datetime.now()
 
 
 @run_async
-def callback_minute(context: CallbackContext):
-    chat_id = context.bot.get_chat(chat_id=context.job.context["chat_name"]).id
+def ban_user(context: CallbackContext):
+    chat_id = context.bot.get_chat(chat_id=context.job.context["chat_id"]).id
     logger.debug(f"get chat.id: {chat_id}")
 
     # sorry about that
     chat_data = context._dispatcher.chat_data.get(chat_id)  # noqa
-
     if chat_data is not None:
-        context.chat_data = {
-            QUARANTINE_STORE_KEY: chat_data[QUARANTINE_STORE_KEY],
-            QUARANTINE_MIN_STORE_KEY: chat_data[QUARANTINE_MIN_STORE_KEY]
-        }
-
-        try:
-            kick_users(chat_id, context)
-        except Exception as e:
-            logger.exception(e)
-        return
-
-    logger.debug("chat_data context is empty, skipped")
-
-
-@run_async
-def kick(chat_id: int, user_id: int, bot: Bot):
-    logger.debug(f"kick user: {user_id}")
-    if not bot.kick_chat_member(chat_id, user_id):
-        logger.info(f"failed to kick user {user_id}")
-
-
-@run_async
-def kick_users(chat_id: int, context: CallbackContext):
-    logger.debug(f"kick users context: {context.chat_data}")
-    bot = context.bot
-    quarantine, quarantine_min = get_quarantine_and_quarantine_min(context)
-    minute = (datetime.datetime.utcnow().minute + 1) % len(quarantine_min)
-    slowpokes = list([x for x in quarantine_min[minute]])
-    for user_id in slowpokes:
-        try:
-            kick(chat_id, user_id, bot)
-            logger.info(f"remove user: {user_id} from quarantine")
-            quarantine_min[minute].remove(user_id)
-            quarantine.pop(user_id)
-        except Exception as e:
-            logger.exception(e)
+        for user in db.find_all_users():
+            if _is_time_gone(user):
+                context.bot.kick_chat_member(chat_id, user['_id'])
+                db.delete_user(user['_id'])
+                logger.debug(f"user banned: {user['_id']}")
